android: Voice messages refactoring (#1511)
* android: Voice messages refactoring

* Different way to block text field from editing while recording voice

* Limited voice record max duration

* Better end of recording when it reaches timeout

* New way of doing things

* Change

* Change

* Stop event refactor

* Stopped state

* Replaced some helpers

* Replaced calls in when()

* Comments

* Change

* Change

+++ apps/android/app/src/main/java/chat/simplex/app/views/helpers/Util.kt
@@ -228,7 +228,7 @@ const val MAX_IMAGE_SIZE_AUTO_RCV: Long = MAX_IMAGE_SIZE * 2
 const val MAX_VOICE_SIZE_AUTO_RCV: Long = MAX_IMAGE_SIZE
 
 const val MAX_VOICE_SIZE_FOR_SENDING: Long = 94680 // 6 chunks * 15780 bytes per chunk
-const val MAX_VOICE_MILLIS_FOR_SENDING: Long = 43_000 // approximately is ok
+const val MAX_VOICE_MILLIS_FOR_SENDING: Int = 43_000
 
 const val MAX_FILE_SIZE: Long = 8000000
